Move RemoteButlerRegistry query methods to base class
Move RemoteButler's registry query implementations to a new base class.  This prepares for using this implementation in DirectButler as well.
