+++ python/lsst/daf/butler/remote_butler/_registry.py
@@ -35,7 +35,6 @@ from lsst.daf.butler import Butler
 from lsst.utils.iteration import ensure_iterable
 
 from .._collection_type import CollectionType
-from .._dataset_association import DatasetAssociation
 from .._dataset_ref import DatasetId, DatasetIdGenEnum, DatasetRef
 from .._dataset_type import DatasetType
 from .._storage_class import StorageClassFactory
@@ -53,29 +52,15 @@ from ..registry import (
     CollectionSummary,
     CollectionTypeError,
     DatasetTypeError,
-    DatasetTypeExpressionError,
-    NoDefaultCollectionError,
-    Registry,
     RegistryDefaults,
 )
-from ..registry._exceptions import ArgumentError
-from ..registry.queries import (
-    ChainedDatasetQueryResults,
-    DataCoordinateQueryResults,
-    DatasetQueryResults,
-    DimensionRecordQueryResults,
-)
-from ..registry.wildcards import CollectionWildcard, DatasetTypeWildcard
+from ..registry._registry_base import RegistryBase
 from ._collection_args import (
     convert_collection_arg_to_glob_string_list,
     convert_dataset_type_arg_to_glob_string_list,
 )
 from ._defaults import DefaultsHolder
 from ._http_connection import RemoteButlerHttpConnection, parse_model
-from .registry._query_common import CommonQueryArguments
-from .registry._query_data_coordinates import QueryDriverDataCoordinateQueryResults
-from .registry._query_datasets import QueryDriverDatasetRefQueryResults
-from .registry._query_dimension_records import QueryDriverDimensionRecordQueryResults
 from .server_models import (
     ExpandDataIdRequestModel,
     ExpandDataIdResponseModel,
@@ -85,7 +70,7 @@ from .server_models import (
 )
 
 
-class RemoteButlerRegistry(Registry):
+class RemoteButlerRegistry(RegistryBase):
     """Implementation of the `Registry` interface for `RemoteButler.
 
     Parameters
@@ -99,7 +84,7 @@ class RemoteButlerRegistry(Registry):
     """
 
     def __init__(self, butler: Butler, defaults: DefaultsHolder, connection: RemoteButlerHttpConnection):
-        self._butler = butler
+        super().__init__(butler)
         self._connection = connection
         self._defaults = defaults
 
@@ -354,170 +339,6 @@ class RemoteButlerRegistry(Registry):
             include_chains=includeChains,
         )
 
-    def queryDatasets(
-        self,
-        datasetType: Any,
-        *,
-        collections: CollectionArgType | None = None,
-        dimensions: Iterable[str] | None = None,
-        dataId: DataId | None = None,
-        where: str = "",
-        findFirst: bool = False,
-        components: bool = False,
-        bind: Mapping[str, Any] | None = None,
-        check: bool = True,
-        **kwargs: Any,
-    ) -> DatasetQueryResults:
-        doomed_by: list[str] = []
-        dimension_group = self.dimensions.conform(dimensions) if dimensions is not None else None
-
-        if collections is None and not self.defaults.collections:
-            raise NoDefaultCollectionError("No collections provided, and no default collections set")
-        if findFirst and collections is not None:
-            wildcard = CollectionWildcard.from_expression(collections)
-            if wildcard.patterns:
-                raise TypeError(
-                    "Collection search patterns not allowed in findFirst search, "
-                    "because collections must be in a specific order."
-                )
-
-        args = self._convert_common_query_arguments(
-            dataId=dataId,
-            where=where,
-            bind=bind,
-            kwargs=kwargs,
-            datasets=None,
-            collections=collections,
-            doomed_by=doomed_by,
-        )
-
-        if not args.collections:
-            doomed_by.append("No datasets can be found because collection list is empty.")
-
-        missing_dataset_types: list[str] = []
-        dataset_types = list(self.queryDatasetTypes(datasetType, missing=missing_dataset_types))
-        if missing_dataset_types:
-            doomed_by.extend(f"Dataset type {name} is not registered." for name in missing_dataset_types)
-
-        if len(dataset_types) == 0:
-            doomed_by.extend(
-                [
-                    f"No registered dataset type matching {t!r} found, so no matching datasets can "
-                    "exist in any collection."
-                    for t in ensure_iterable(datasetType)
-                ]
-            )
-            return ChainedDatasetQueryResults([], doomed_by=doomed_by)
-
-        query_results = [
-            QueryDriverDatasetRefQueryResults(
-                self._butler.query,
-                args,
-                dataset_type=dt,
-                find_first=findFirst,
-                extra_dimensions=dimension_group,
-                doomed_by=doomed_by,
-                expanded=False,
-            )
-            for dt in dataset_types
-        ]
-        if len(query_results) == 1:
-            return query_results[0]
-        else:
-            return ChainedDatasetQueryResults(query_results)
-
-    def queryDataIds(
-        self,
-        dimensions: DimensionGroup | Iterable[str] | str,
-        *,
-        dataId: DataId | None = None,
-        datasets: Any = None,
-        collections: CollectionArgType | None = None,
-        where: str = "",
-        components: bool = False,
-        bind: Mapping[str, Any] | None = None,
-        check: bool = True,
-        **kwargs: Any,
-    ) -> DataCoordinateQueryResults:
-        if collections is not None and datasets is None:
-            raise ArgumentError(f"Cannot pass 'collections' (='{collections}') without 'datasets'.")
-
-        dimensions = self.dimensions.conform(dimensions)
-        args = self._convert_common_query_arguments(
-            dataId=dataId, where=where, bind=bind, kwargs=kwargs, datasets=datasets, collections=collections
-        )
-        return QueryDriverDataCoordinateQueryResults(
-            self._butler.query, dimensions=dimensions, expanded=False, args=args
-        )
-
-    def queryDimensionRecords(
-        self,
-        element: DimensionElement | str,
-        *,
-        dataId: DataId | None = None,
-        datasets: Any = None,
-        collections: CollectionArgType | None = None,
-        where: str = "",
-        components: bool = False,
-        bind: Mapping[str, Any] | None = None,
-        check: bool = True,
-        **kwargs: Any,
-    ) -> DimensionRecordQueryResults:
-        if not isinstance(element, DimensionElement):
-            element = self.dimensions.elements[element]
-
-        args = self._convert_common_query_arguments(
-            dataId=dataId, where=where, bind=bind, kwargs=kwargs, datasets=datasets, collections=collections
-        )
-
-        return QueryDriverDimensionRecordQueryResults(self._butler.query, element, args)
-
-    def _convert_common_query_arguments(
-        self,
-        *,
-        dataId: DataId | None = None,
-        datasets: object | None = None,
-        collections: CollectionArgType | None = None,
-        where: str = "",
-        bind: Mapping[str, Any] | None = None,
-        kwargs: dict[str, int | str],
-        doomed_by: list[str] | None = None,
-    ) -> CommonQueryArguments:
-        dataset_types = self._resolve_dataset_types(datasets)
-        if dataset_types and collections is None and not self.defaults.collections:
-            raise NoDefaultCollectionError("'collections' must be provided if 'datasets' is provided")
-        return CommonQueryArguments(
-            dataId=dataId,
-            where=where,
-            bind=dict(bind) if bind else None,
-            kwargs=dict(kwargs),
-            dataset_types=dataset_types,
-            collections=self._resolve_collections(collections, doomed_by),
-        )
-
-    def queryDatasetAssociations(
-        self,
-        datasetType: str | DatasetType,
-        collections: CollectionArgType | None = ...,
-        *,
-        collectionTypes: Iterable[CollectionType] = CollectionType.all(),
-        flattenChains: bool = False,
-    ) -> Iterator[DatasetAssociation]:
-        # queryCollections only accepts DatasetType.
-        if isinstance(datasetType, str):
-            datasetType = self.getDatasetType(datasetType)
-        resolved_collections = self.queryCollections(
-            collections, datasetType=datasetType, collectionTypes=collectionTypes, flattenChains=flattenChains
-        )
-        with self._butler.query() as query:
-            query = query.join_dataset_search(datasetType, resolved_collections)
-            result = query.general(
-                datasetType.dimensions,
-                dataset_fields={datasetType.name: {"dataset_id", "run", "collection", "timespan"}},
-                find_first=False,
-            )
-            yield from DatasetAssociation.from_query_result(result, datasetType)
-
     @property
     def storageClasses(self) -> StorageClassFactory:
         return self._butler.storageClasses
@@ -525,39 +346,6 @@ class RemoteButlerRegistry(Registry):
     @storageClasses.setter
     def storageClasses(self, value: StorageClassFactory) -> None:
         raise NotImplementedError()
-
-    def _resolve_collections(
-        self, collections: CollectionArgType | None, doomed_by: list[str] | None = None
-    ) -> list[str] | None:
-        if collections is None:
-            return list(self.defaults.collections)
-
-        wildcard = CollectionWildcard.from_expression(collections)
-        if wildcard.patterns:
-            result = list(self.queryCollections(collections))
-            if not result and doomed_by is not None:
-                doomed_by.append(f"No collections found matching expression {wildcard}")
-            return result
-        else:
-            return list(wildcard.strings)
-
-    def _resolve_dataset_types(self, dataset_types: object | None) -> list[str]:
-        if dataset_types is None:
-            return []
-
-        if dataset_types is ...:
-            raise TypeError(
-                "'...' not permitted for 'datasets'"
-                " -- searching for all dataset types does not constrain the search."
-            )
-
-        wildcard = DatasetTypeWildcard.from_expression(dataset_types)
-        if wildcard.patterns:
-            raise DatasetTypeExpressionError(
-                "Dataset type wildcard expressions are not supported in this context."
-            )
-        else:
-            return list(wildcard.values.keys())
 
 
 def _is_component_dataset_type(dataset_type: DatasetType | str) -> bool:
